# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API de Gestión de Eventos",
        "description": "Documentación de la API para la gestión de eventos y reservas.",
        "contact": {
            "email": "tu_correo@example.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/events/{event}/reservations": {
            "post": {
                "tags": [
                    "Reservations"
                ],
                "summary": "Crear una nueva reserva para un evento",
                "description": "Crea una nueva reserva para un evento específico. Requiere autenticación.",
                "operationId": "storeReservation",
                "parameters": [
                    {
                        "name": "event",
                        "in": "path",
                        "description": "ID del evento para el cual se crea la reserva",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "user_name",
                                    "user_email",
                                    "seats"
                                ],
                                "properties": {
                                    "user_name": {
                                        "description": "Nombre del usuario que realiza la reserva",
                                        "type": "string",
                                        "maxLength": 255
                                    },
                                    "user_email": {
                                        "description": "Correo electrónico del usuario",
                                        "type": "string",
                                        "format": "email",
                                        "maxLength": 255
                                    },
                                    "seats": {
                                        "description": "Número de asientos reservados",
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Reserva creada exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Reservation"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Solicitud incorrecta"
                    },
                    "401": {
                        "description": "No autenticado"
                    },
                    "404": {
                        "description": "Evento no encontrado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/reservations/{reservation}": {
            "delete": {
                "tags": [
                    "Reservations"
                ],
                "summary": "Eliminar una reserva existente",
                "description": "Elimina una reserva existente basada en su ID. Requiere autenticación.",
                "operationId": "deleteReservation",
                "parameters": [
                    {
                        "name": "reservation",
                        "in": "path",
                        "description": "ID de la reserva a eliminar",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Reserva eliminada exitosamente"
                    },
                    "401": {
                        "description": "No autenticado"
                    },
                    "404": {
                        "description": "Reserva no encontrada"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
...
            "Reservation": {
                "title": "Reservation",
                "description": "Reservation model",
                "properties": {
                    "id": {
                        "title": "ID",
                        "description": "Reservation ID",
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "event_id": {
                        "title": "Event ID",
                        "description": "ID of the event the reservation belongs to",
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "user_name": {
                        "title": "User Name",
                        "description": "Name of the user making the reservation",
                        "type": "string",
                        "example": "John Doe"
                    },
                    "user_email": {
                        "title": "User Email",
                        "description": "Email of the user making the reservation",
                        "type": "string",
                        "format": "email",
                        "example": "john.doe@example.com"
                    },
                    "seats": {
                        "title": "Seats",
                        "description": "Number of seats reserved",
                        "type": "integer",
                        "example": 2
                    },
                    "created_at": {
                        "title": "Created At",
                        "description": "Timestamp of when the reservation was created",
                        "type": "string",
                        "format": "date-time",
                        "example": "2023-01-01 12:00:00"
                    },
                    "updated_at": {
                        "title": "Updated At",
                        "description": "Timestamp of when the reservation was last updated",
                        "type": "string",
                        "format": "date-time",
                        "example": "2023-01-01 12:00:00"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "Reservation"
                }
            }
        }
    },
...
}